+++ src/hoi4_world/states/hoi4_state.h
@@ -79,7 +79,7 @@ class State
    [[nodiscard]] int GetAirBaseLevel() const { return air_base_level_; }
    [[nodiscard]] std::set<std::string> GetCores() const { return cores_; }
    [[nodiscard]] float GetVic3Infrastructure() const { return vic3_infrastructure_; }
-   [[nodiscard]] float GetInfrastructure() const { return infrastructure_; }
+   [[nodiscard]] int GetInfrastructure() const { return infrastructure_; }
 
    void IncreaseAirBaseLevel(int amount) { air_base_level_ = std::min(air_base_level_ + amount, 10); }
